moved and renamed get_img_cover
